Expanded counters from int to long long.

+++ h/visitors/IntegerPartitionVisitorCounter.h
@@ -8,7 +8,7 @@ public:
     void visit(IntegerPartitionsGenerator::Partition& partition, std::ostream* partitionOut) override;
     void results(std::ostream* resultsOut) override;
 private:
-    int count = 0;
+    unsigned long long count = 0;
 };
 
 #endif //PARTITIONSGENERATION_INTEGERPARTITIONVISITORCOUNTER_H
